requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-ppc64le
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu_881.g2babef4
absl-py=0.7.1=py36_0
aiohttp=3.7.3=py36h140841e_1
argon2-cffi=20.1.0=py36h7b6447c_1
astor=0.7.1=py36_0
async-timeout=3.0.1=py36_0
async_generator=1.10=py36h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=openblas
bleach=3.2.1=py_0
blosc=1.20.1=hd408876_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py36h140841e_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.16.1=h7b6447c_0
ca-certificates=2020.12.8=h6ffa863_0
cachetools=4.2.0=pyhd3eb1b0_0
certifi=2020.12.5=py36h6ffa863_0
cffi=1.14.4=py36hf9d8e4b_0
chardet=3.0.4=py36h6ffa863_1003
charls=2.1.0=he6710b0_2
cloudpickle=1.6.0=py_0
cryptography=3.3.1=py36hb567c45_0
cudatoolkit=10.1.243=616.gc122b8b
cudnn=7.6.3_10.1=590.g5627c5e
cycler=0.10.0=py36_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2020.12.0=pyhd3eb1b0_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
freetype=2.10.4=h5ab3b9f_0
gast=0.2.2=py36_0
giflib=5.1.4=h14c3975_1
gmp=6.1.2=h7f7056e_2
google-api-core=1.22.2=py36_0
google-auth=1.23.0=pyhd3eb1b0_0
google-cloud-core=1.4.4=pyhd3eb1b0_0
google-cloud-storage=1.33.0=pyhd3eb1b0_0
google-crc32c=1.0.0=py36h7b6447c_0
google-pasta=0.1.6=py36_584.g6a97e08
google-resumable-media=1.1.0=py_1
googleapis-common-protos=1.52.0=py36_0
graphsurgeon=0.4.1=py36_634.g4ac44fb
grpcio=1.16.1=py36hf8bcb03_1
h5py=2.10.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
idna=2.10=py_0
idna_ssl=1.1.0=py36_0
imagecodecs=2020.5.30=py36ha36bdba_2
imageio=2.9.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.5.1=py_1
jedi=0.17.0=py36_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=hcb7ba68_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py36h6ffa863_0
jupyterlab_pygments=0.1.2=py_0
jxrlib=1.1=h7b6447c_2
keras=2.1.6=py36_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.0=py36h29c3540_0
lcms2=2.11=h396b838_0
ld_impl_linux-ppc64le=2.33.1=h0f24833_7
libaec=1.0.4=he6710b0_1
libcrc32c=1.1.1=he6710b0_2
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=8.2.0=h822a55f_1
libgfortran-ng=7.3.0=h822a55f_1
libopenblas=0.3.10=h5a2b251_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=hd408876_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=8.2.0=h822a55f_1
libtiff=4.1.0=h2733197_1
libwebp=1.0.1=h8e7db2f_0
libzopfli=1.0.3=he6710b0_0
lz4-c=1.9.2=h39c303f_3
markdown=3.3.2=py36_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py36h76a9e4f_0
mistune=0.8.4=py36h7b6447c_0
multidict=4.7.6=py36h7b6447c_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py36_0
nbformat=5.0.8=py_0
nccl=2.4.8=586.gdba67b7
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
networkx=2.5=py_0
notebook=6.1.4=py36_0
numpy=1.16.6=py36h30dfecb_0
numpy-base=1.16.6=py36h2f8d375_0
olefile=0.46=py36_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1i=h140841e_0
packaging=20.7=pyhd3eb1b0_0
pandas=1.1.3=py36hdf5156a_0
pandoc=2.11=h7557452_0
pandocfilters=1.4.3=py36h6ffa863_1
parso=0.8.1=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=7.1.2=py36haac5956_0
pip=20.3.1=py36h6ffa863_0
powerai-license=1.6.2=734.g398a206
powerai-release=1.6.2=590.g8b407ea
powerai-tools=1.6.2=584.ga57254d
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.8.0=py36he6710b0_0
ptyprocess=0.6.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.3=pyhd3eb1b0_0
pyopenssl=20.0.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyrsistent=0.17.3=py36h7b6447c_0
pysocks=1.7.1=py36h6ffa863_0
python=3.6.12=ha7b6439_2
python-dateutil=2.8.1=py_0
pytz=2020.4=pyhd3eb1b0_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.3.1=py36h7b6447c_1
pyzmq=20.0.0=py36h29c3540_1
readline=8.0=h7b6447c_0
requests=2.25.0=pyhd3eb1b0_0
rsa=4.6=py_0
scikit-image=0.17.2=py36hdf5156a_0
scikit-learn=0.23.2=py36h0573a6f_0
scipy=1.5.2=py36habc2bb6_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=51.0.0=py36h6ffa863_2
six=1.12.0=py36_0
snappy=1.1.8=he6710b0_0
sqlite=3.33.0=hbc83047_0
tensorboard=1.15.0=pyhb230dea_0
tensorflow=1.15.4=gpu_py36_881.g2babef4
tensorflow-base=1.15.4=gpu_py36_df8c55c_64440.g07ecaec
tensorflow-estimator=1.15.1=pyh2649769_0
tensorflow-gpu=1.15.4=881.g2babef4
tensorrt=6.0.1.5=py36_634.g4ac44fb
termcolor=1.1.0=py36_1
terminado=0.9.1=py36_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.12.8=pyhd3eb1b0_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
tornado=6.1=py36h140841e_0
tqdm=4.54.1=pyhd3eb1b0_0
traitlets=4.3.3=py36_0
typing-extensions=3.7.2=py36_1596.g5249206
typing_extensions=3.7.2=py36_1596.g5249206
uff=0.6.5=py36_634.g4ac44fb
urllib3=1.25.11=py_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
werkzeug=1.0.1=py_0
wheel=0.36.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.5.1=py36h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.3.1=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
